+++ CHANGELOG.md
@@ -48,6 +48,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed bugs
 
+* Conflicts involving non-empty files that do not end in a newline no longer
+  look broken when materialized.
+  [#3968](https://github.com/martinvonz/jj/issues/3968). This is a partial fix,
+  see also [#3975](https://github.com/martinvonz/jj/issues/3968) which is not
+  yet fixed.
+
 ## [0.19.0] - 2024-07-03
 
 ### Breaking changes